Refine selection module boundaries
