+++ scripts/internal/check/allowlists/file_size_budget_allowlist.txt
@@ -14,16 +14,6 @@
 # tests. Split only if concrete ownership or correctness pressure appears.
 src/app_core/actions/catalog/kinds.rs
 
-# Playback selection drag/retarget logic and its local unit tests currently form
-# one cohesive transport subdomain. Revisit only if a clearer module boundary
-# emerges than the current gesture/playback coupling.
-src/app/controller/playback/transport/selection.rs
-
-# Selection range geometry, fade parameters, and gain evaluation intentionally
-# stay together as one waveform-editing domain model. Revisit only if a
-# genuinely separate subdomain emerges beyond the current cohesion note.
-src/selection/range.rs
-
 # Vendored Radiant still contains a small set of oversized legacy modules and
 # contract-heavy test files. Keep them explicit so the repaired budget script
 # can enforce everything else in `vendor/radiant/src`.
