+++ src/selection/mod.rs
@@ -2,12 +2,17 @@
 //! This module keeps selection math pure and testable so the UI integration code can stay small.
 
 /// Selection range geometry, fades, and gain helpers.
+mod fade;
+/// Decoded frame conversion helpers for normalized selection bounds.
+mod frames;
 mod range;
 /// Selection drag state machine and edge-drag behavior.
 mod state;
 
-pub(crate) use range::fade_gain_at_position;
-pub use range::{FadeParams, SampleFrameRange, SelectionRange};
+pub use fade::FadeParams;
+pub(crate) use fade::fade_gain_at_position;
+pub use frames::SampleFrameRange;
+pub use range::SelectionRange;
 pub use state::{SelectionEdge, SelectionState};
 
 #[cfg(test)]
